pyproject.toml
[project]
name = "ai-minds"
version = "1.0.0"
description = "AI MINDS - Multimodal Personal Knowledge Assistant"
requires-python = ">=3.9"
dependencies = [
    # Core ML & Embeddings
    "torch>=2.0.0",
    "sentence-transformers>=2.2.0",
    "transformers>=4.30.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
    
    # Data Collection
    "browser-history>=0.3.2",
    "pyperclip>=1.8.2",
    "Pillow>=10.0.0",
    "pywin32>=306 ; sys_platform == 'win32'",
    
    # Google APIs (Calendar & Email)
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

# Optional AI features
ai-full = [
    "openai-whisper>=20230314",
    "pytesseract>=0.3.10",
    "easyocr>=1.7.0",
    "pypdf2>=3.0.0",
    "python-docx>=1.0.0",
    "spacy>=3.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["Data_Layer", "Core", "API", "Tests"]

[tool.uv]
dev-dependencies = []

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.black]
line-length = 100
target-version = ["py39"]
